refactor(backend): extract tag operations to TagRepository
